Split native Zenmap profile helper views
